refactor: isolate file-native turn generation
