+++ crates/agent-engine/src/llm/input_projection.rs
@@ -1,6 +1,6 @@
-use super::{GenerationRequest, InstructionRole, Message, MessageRole, ToolCall};
+use super::{Message, MessageRole, ToolCall};
 use crate::agent_machine::{Message as MachineMessage, MessageRole as MachineMessageRole};
-use crate::tape::{ContentPart, parts_to_text};
+use crate::tape::ContentPart;
 
 const MAX_PROJECTED_TOOL_PAYLOAD_SIZE: usize = 30_000;
 const PROJECTION_TRUNCATION_MARKER: &str = "...[truncated]";
@@ -65,29 +65,6 @@ pub(crate) fn project_messages(
             }
         })
         .collect()
-}
-
-pub(crate) fn with_provider_input(
-    request: GenerationRequest,
-    instruction_role: InstructionRole,
-    messages: &[MachineMessage],
-) -> GenerationRequest {
-    let (key, projected) = match instruction_role {
-        InstructionRole::ResponsesInstructions => (
-            "responses_input_items",
-            build_responses_input_items(messages),
-        ),
-        InstructionRole::Developer => (
-            "chat_completions_messages",
-            build_chat_completions_messages(messages),
-        ),
-        InstructionRole::AnthropicSystem => {
-            ("anthropic_messages", build_anthropic_messages(messages))
-        }
-        InstructionRole::System => return request,
-    };
-
-    request.with_extra_param(key, serde_json::Value::Array(projected))
 }
 
 fn non_empty_trimmed(value: &str) -> Option<String> {
@@ -241,455 +218,6 @@ fn utf8_prefix(text: &str, max_len: usize) -> String {
     }
 
     text[..end].to_string()
-}
-
-fn responses_attachment_input_part(
-    hash: &str,
-    mime_type: &str,
-    metadata: &serde_json::Value,
-) -> serde_json::Value {
-    if mime_type.starts_with("image/") {
-        if let Some(image_url) = metadata
-            .get("image_url")
-            .or_else(|| metadata.get("file_url"))
-            .or_else(|| metadata.get("url"))
-            .and_then(serde_json::Value::as_str)
-        {
-            return serde_json::json!({
-                "type": "input_image",
-                "image_url": image_url,
-            });
-        }
-        if let Some(file_id) = metadata.get("file_id").and_then(serde_json::Value::as_str) {
-            return serde_json::json!({
-                "type": "input_image",
-                "file_id": file_id,
-            });
-        }
-    }
-
-    if let Some(file_id) = metadata.get("file_id").and_then(serde_json::Value::as_str) {
-        return serde_json::json!({
-            "type": "input_file",
-            "file_id": file_id,
-        });
-    }
-    if let Some(file_url) = metadata
-        .get("file_url")
-        .or_else(|| metadata.get("url"))
-        .and_then(serde_json::Value::as_str)
-    {
-        return serde_json::json!({
-            "type": "input_file",
-            "file_url": file_url,
-        });
-    }
-
-    serde_json::json!({
-        "type": "input_text",
-        "text": format!("[attachment: {hash} ({mime_type})]"),
-    })
-}
-
-fn chat_completions_attachment_content_part(
-    hash: &str,
-    mime_type: &str,
-    metadata: &serde_json::Value,
-) -> serde_json::Value {
-    if mime_type.starts_with("image/")
-        && let Some(image_url) = metadata
-            .get("image_url")
-            .or_else(|| metadata.get("file_url"))
-            .or_else(|| metadata.get("url"))
-            .and_then(serde_json::Value::as_str)
-    {
-        return serde_json::json!({
-            "type": "image_url",
-            "image_url": { "url": image_url },
-        });
-    }
-    if let Some(file_id) = metadata.get("file_id").and_then(serde_json::Value::as_str) {
-        return serde_json::json!({
-            "type": "file",
-            "file": { "file_id": file_id },
-        });
-    }
-
-    serde_json::json!({
-        "type": "text",
-        "text": format!("[attachment: {hash} ({mime_type})]"),
-    })
-}
-
-fn anthropic_attachment_content_block(
-    hash: &str,
-    mime_type: &str,
-    metadata: &serde_json::Value,
-) -> serde_json::Value {
-    let block_type = if mime_type.starts_with("image/") {
-        "image"
-    } else {
-        "document"
-    };
-
-    if let Some(file_id) = metadata.get("file_id").and_then(serde_json::Value::as_str) {
-        let mut block = serde_json::json!({
-            "type": block_type,
-            "source": {
-                "type": "file",
-                "file_id": file_id,
-            },
-        });
-        add_anthropic_document_title(&mut block, block_type, metadata);
-        return block;
-    }
-    if let Some(url) = metadata
-        .get("file_url")
-        .or_else(|| metadata.get("image_url"))
-        .or_else(|| metadata.get("url"))
-        .and_then(serde_json::Value::as_str)
-    {
-        let mut block = serde_json::json!({
-            "type": block_type,
-            "source": {
-                "type": "url",
-                "url": url,
-            },
-        });
-        add_anthropic_document_title(&mut block, block_type, metadata);
-        return block;
-    }
-
-    serde_json::json!({
-        "type": "text",
-        "text": format!("[attachment: {hash} ({mime_type})]"),
-    })
-}
-
-fn add_anthropic_document_title(
-    block: &mut serde_json::Value,
-    block_type: &str,
-    metadata: &serde_json::Value,
-) {
-    if block_type == "document"
-        && let Some(title) = metadata.get("title").and_then(serde_json::Value::as_str)
-    {
-        block["title"] = serde_json::Value::String(title.to_string());
-    }
-}
-
-fn responses_message_content(parts: &[ContentPart]) -> Option<serde_json::Value> {
-    let needs_array = parts.iter().any(|part| {
-        !matches!(
-            part,
-            ContentPart::Text { .. } | ContentPart::Thinking { .. }
-        )
-    });
-    if !needs_array {
-        let text = parts_to_text(parts);
-        return (!text.trim().is_empty()).then_some(serde_json::Value::String(text));
-    }
-
-    let content_parts = parts
-        .iter()
-        .filter_map(|part| match part {
-            ContentPart::Text { text } if !text.trim().is_empty() => Some(serde_json::json!({
-                "type": "input_text",
-                "text": text,
-            })),
-            ContentPart::Attachment {
-                hash,
-                mime_type,
-                metadata,
-            } => Some(responses_attachment_input_part(hash, mime_type, metadata)),
-            ContentPart::Structured { data } => Some(serde_json::json!({
-                "type": "input_text",
-                "text": data.to_string(),
-            })),
-            _ => None,
-        })
-        .collect::<Vec<_>>();
-
-    (!content_parts.is_empty()).then_some(serde_json::Value::Array(content_parts))
-}
-
-fn chat_completions_message_content(parts: &[ContentPart]) -> Option<serde_json::Value> {
-    let needs_array = parts.iter().any(|part| {
-        !matches!(
-            part,
-            ContentPart::Text { .. } | ContentPart::Thinking { .. }
-        )
-    });
-    if !needs_array {
-        let text = parts_to_text(parts);
-        return (!text.trim().is_empty()).then_some(serde_json::Value::String(text));
-    }
-
-    let content_parts = parts
-        .iter()
-        .filter_map(|part| match part {
-            ContentPart::Text { text } if !text.trim().is_empty() => Some(serde_json::json!({
-                "type": "text",
-                "text": text,
-            })),
-            ContentPart::Attachment {
-                hash,
-                mime_type,
-                metadata,
-            } => Some(chat_completions_attachment_content_part(
-                hash, mime_type, metadata,
-            )),
-            ContentPart::Structured { data } => Some(serde_json::json!({
-                "type": "text",
-                "text": data.to_string(),
-            })),
-            _ => None,
-        })
-        .collect::<Vec<_>>();
-
-    (!content_parts.is_empty()).then_some(serde_json::Value::Array(content_parts))
-}
-
-fn anthropic_message_content(parts: &[ContentPart]) -> Vec<serde_json::Value> {
-    parts
-        .iter()
-        .filter_map(|part| match part {
-            ContentPart::Text { text } if !text.trim().is_empty() => Some(serde_json::json!({
-                "type": "text",
-                "text": text,
-            })),
-            ContentPart::Thinking { text, signature } if !text.trim().is_empty() => {
-                let mut block = serde_json::json!({
-                    "type": "thinking",
-                    "thinking": text,
-                });
-                if let Some(signature) = signature
-                    .as_deref()
-                    .filter(|value| !value.trim().is_empty())
-                {
-                    block["signature"] = serde_json::Value::String(signature.to_string());
-                }
-                Some(block)
-            }
-            ContentPart::RedactedThinking { data } if !data.trim().is_empty() => {
-                Some(serde_json::json!({
-                    "type": "redacted_thinking",
-                    "data": data,
-                }))
-            }
-            ContentPart::Attachment {
-                hash,
-                mime_type,
-                metadata,
-            } => Some(anthropic_attachment_content_block(
-                hash, mime_type, metadata,
-            )),
-            ContentPart::Structured { data } => Some(serde_json::json!({
-                "type": "text",
-                "text": data.to_string(),
-            })),
-            _ => None,
-        })
-        .collect()
-}
-
-fn build_responses_input_items(messages: &[MachineMessage]) -> Vec<serde_json::Value> {
-    let mut input = Vec::new();
-
-    for message in messages {
-        match message {
-            MachineMessage::Tool { responses } => {
-                for response in responses {
-                    input.push(serde_json::json!({
-                        "type": "function_call_output",
-                        "call_id": response.id,
-                        "output": project_tool_response(&response.content),
-                    }));
-                }
-            }
-            MachineMessage::Assistant {
-                parts,
-                tool_requests,
-            } => {
-                if let Some(signature) = message.thinking_signature() {
-                    input.push(serde_json::json!({
-                        "type": "reasoning",
-                        "encrypted_content": signature,
-                    }));
-                }
-                if let Some(content) = responses_message_content(parts) {
-                    input.push(serde_json::json!({
-                        "role": "assistant",
-                        "content": content,
-                    }));
-                }
-                for tool_request in tool_requests {
-                    input.push(serde_json::json!({
-                        "type": "function_call",
-                        "call_id": tool_request.id,
-                        "name": tool_request.name,
-                        "arguments": tool_request.arguments.to_string(),
-                    }));
-                }
-            }
-            MachineMessage::User { parts }
-            | MachineMessage::System { parts }
-            | MachineMessage::Context { parts } => {
-                if let Some(content) = responses_message_content(parts) {
-                    let role = if message.role() == MachineMessageRole::User {
-                        "user"
-                    } else {
-                        "developer"
-                    };
-                    input.push(serde_json::json!({
-                        "role": role,
-                        "content": content,
-                    }));
-                }
-            }
-        }
-    }
-
-    input
-}
-
-fn build_chat_completions_messages(messages: &[MachineMessage]) -> Vec<serde_json::Value> {
-    let mut projected = Vec::new();
-
-    for message in messages {
-        match message {
-            MachineMessage::Tool { responses } => {
-                for response in responses {
-                    projected.push(serde_json::json!({
-                        "role": "tool",
-                        "content": project_tool_response(&response.content),
-                        "tool_call_id": response.id,
-                    }));
-                }
-            }
-            MachineMessage::Assistant {
-                parts,
-                tool_requests,
-            } => {
-                let mut message_value = serde_json::json!({ "role": "assistant" });
-                if let Some(content) = chat_completions_message_content(parts) {
-                    message_value["content"] = content;
-                }
-                if let Some(thinking) = message.thinking_content() {
-                    message_value["reasoning_content"] = serde_json::Value::String(thinking);
-                }
-                if let Some(signature) = message.thinking_signature() {
-                    message_value["reasoning"] = serde_json::json!({
-                        "encrypted_content": signature,
-                    });
-                }
-                if !tool_requests.is_empty() {
-                    message_value["tool_calls"] = serde_json::Value::Array(
-                        tool_requests
-                            .iter()
-                            .map(|tool_request| {
-                                serde_json::json!({
-                                    "id": tool_request.id,
-                                    "type": "function",
-                                    "function": {
-                                        "name": tool_request.name,
-                                        "arguments": tool_request.arguments.to_string(),
-                                    },
-                                })
-                            })
-                            .collect(),
-                    );
-                }
-                projected.push(message_value);
-            }
-            MachineMessage::User { parts } => {
-                if let Some(content) = chat_completions_message_content(parts) {
-                    projected.push(serde_json::json!({
-                        "role": "user",
-                        "content": content,
-                    }));
-                }
-            }
-            MachineMessage::System { parts } | MachineMessage::Context { parts } => {
-                if let Some(content) = chat_completions_message_content(parts) {
-                    projected.push(serde_json::json!({
-                        "role": "developer",
-                        "content": content,
-                    }));
-                }
-            }
-        }
-    }
-
-    projected
-}
-
-fn build_anthropic_messages(messages: &[MachineMessage]) -> Vec<serde_json::Value> {
-    let mut projected = Vec::new();
-    let mut known_tool_use_ids = std::collections::HashSet::new();
-
-    for message in messages {
-        match message {
-            MachineMessage::Tool { responses } => {
-                for response in responses {
-                    let content = project_tool_response(&response.content);
-                    let block = if known_tool_use_ids.contains(&response.id) {
-                        Some(serde_json::json!({
-                            "type": "tool_result",
-                            "tool_use_id": response.id,
-                            "content": content,
-                        }))
-                    } else if !content.trim().is_empty() {
-                        Some(serde_json::json!({
-                            "type": "text",
-                            "text": content,
-                        }))
-                    } else {
-                        None
-                    };
-                    if let Some(block) = block {
-                        projected.push(serde_json::json!({
-                            "role": "user",
-                            "content": [block],
-                        }));
-                    }
-                }
-            }
-            MachineMessage::Assistant {
-                parts,
-                tool_requests,
-            } => {
-                let mut blocks = anthropic_message_content(parts);
-                for tool_request in tool_requests {
-                    known_tool_use_ids.insert(tool_request.id.clone());
-                    blocks.push(serde_json::json!({
-                        "type": "tool_use",
-                        "id": tool_request.id,
-                        "name": tool_request.name,
-                        "input": tool_request.arguments,
-                    }));
-                }
-                if !blocks.is_empty() {
-                    projected.push(serde_json::json!({
-                        "role": "assistant",
-                        "content": blocks,
-                    }));
-                }
-            }
-            MachineMessage::User { parts } => {
-                let blocks = anthropic_message_content(parts);
-                if !blocks.is_empty() {
-                    projected.push(serde_json::json!({
-                        "role": "user",
-                        "content": blocks,
-                    }));
-                }
-            }
-            MachineMessage::System { .. } | MachineMessage::Context { .. } => {}
-        }
-    }
-
-    projected
 }
 
 #[cfg(test)]
